+++ src/genh5_abstractdataset.cpp
@@ -12,16 +12,53 @@
 
 GenH5::AbstractDataSet::AbstractDataSet() = default;
 
+namespace GenH5
+{
+
+/// We want to check if the other datatype is a compound with compatible
+/// memory layout but has default member names. In this case we should use the
+/// datatype of the dataset so that the correct member names are used
+inline bool isDefaultCompoundType(GenH5::DataType dtype, GenH5::DataType other)
+{
+    // We have to check this only for compound types
+    if (!dtype.isCompound() || !other.isCompound())
+    {
+        return false;
+    }
+
+    CompoundMembers const tMembers = dtype.compoundMembers();
+    CompoundMembers const oMembers = other.compoundMembers();
+    int size = tMembers.size();
+
+    // Compound size must match
+    if (size != oMembers.size())
+    {
+        return false;
+    }
+
+    for (int i = 0; i < size; ++i)
+    {
+        if (// the offset must be the same (identical memory layout)
+            tMembers[i].offset != oMembers[i].offset ||
+            // the meber type must be the same
+            tMembers[i].type   != oMembers[i].type ||
+            // names should not match
+            tMembers[i].name   == oMembers[i].name ||
+            // we expect a default compound type name
+            oMembers[i].name != String{"type_"} + QString::number(i))
+        {
+            return false;
+        }
+    }
+
+    return true;
+}
+
+} // namespace
+
 bool
 GenH5::AbstractDataSet::write(void const* data, Optional<DataType> dtype) const
 {
-    auto space = dataSpace();
-    if (space.isNull())
-    {
-        qCritical() << "HDF5: Writing data vector failed!"
-                    << "(Null dataspace)";
-        return false;
-    }
     if (!data)
     {
         qCritical() << "HDF5: Writing data vector failed!"
@@ -29,8 +66,17 @@ GenH5::AbstractDataSet::write(void const* data, Optional<DataType> dtype) const
         return false;
     }
 
+    auto space = dataSpace();
+    if (space.isNull())
+    {
+        qCritical() << "HDF5: Writing data vector failed!"
+                    << "(Null dataspace)";
+        return false;
+    }
+
     auto type = dataType();
-    if (dtype.isDefault() || dtype == type)
+    // use the datatype of the dataset if not specified
+    if (dtype.isDefault() || isDefaultCompoundType(type, dtype))
     {
         dtype = type;
     }
@@ -41,6 +87,12 @@ GenH5::AbstractDataSet::write(void const* data, Optional<DataType> dtype) const
 bool
 GenH5::AbstractDataSet::read(void* data, Optional<DataType> dtype) const
 {
+    if (!data)
+    {
+        qCritical() << "HDF5: Reading data vector failed!"
+                    << "(Data vector is invalid)";
+        return false;
+    }
 
     auto space = dataSpace();
     if (space.isNull())
@@ -49,15 +101,10 @@ GenH5::AbstractDataSet::read(void* data, Optional<DataType> dtype) const
                     << "(Null dataspace)";
         return false;
     }
-    if (!data)
-    {
-        qCritical() << "HDF5: Reading data vector failed!"
-                    << "(Data vector is invalid)";
-        return false;
-    }
 
     auto type = dataType();
-    if (dtype.isDefault() || dtype == type)
+    // use the datatype of the dataset if not specified
+    if (dtype.isDefault() || isDefaultCompoundType(type, dtype))
     {
         dtype = type;
     }