+++ src/genh5_dataset.h
@@ -30,11 +30,14 @@ public:
     DataSet();
     explicit DataSet(std::shared_ptr<File> file, H5::DataSet dset);
 
+#ifndef GENH5_NO_DEPRECATED_SYMBOLS
     /**
      * @brief allows access of the base hdf5 object
      * @return base hdf5 object
      */
+    [[deprecated("use id() instead")]]
     H5::DataSet const& toH5() const noexcept;
+#endif
 
     /**
      * @brief id or handle of the hdf5 resource
@@ -51,6 +54,7 @@ public:
      * @brief properties used to create this object.
      * @return create properties
      */
+    [[deprecated("use cProperties() instead")]]
     DataSetCProperties properties() const noexcept(false)
     {
         return cProperties();