+++ src/genh5_datasetcproperties.h
@@ -34,11 +34,14 @@ public:
     explicit DataSetCProperties(Dimensions const& chunkDimensions,
                                 int compression = 0) noexcept(false);
 
+#ifndef GENH5_NO_DEPRECATED_SYMBOLS
     /**
      * @brief allows access of the base hdf5 object
      * @return base hdf5 object
      */
+    [[deprecated("use id() instead")]]
     H5::DSetCreatPropList const& toH5() const noexcept;
+#endif
 
     /**
      * @brief id or handle of the hdf5 resource