+++ src/genh5_file.h
@@ -67,11 +67,14 @@ public:
     explicit File(H5::H5File file);
     explicit File(String path, FileAccessFlags flags = ReadWrite);
 
+#ifndef GENH5_NO_DEPRECATED_SYMBOLS
     /**
      * @brief allows access of the base hdf5 object.
      * @return base hdf5 object
      */
+    [[deprecated("use id() instead")]]
     H5::H5File const& toH5() const noexcept;
+#endif
 
     /**
      * @brief id or handle of the hdf5 resource.